+++ modules/azurerm_route_table/variables.tf
@@ -59,12 +59,6 @@ variable "routes" {
   }
 }
 
-# Subnet Associations
-variable "subnet_ids_to_associate" {
-  description = "A list of subnet IDs to associate with this route table."
-  type        = list(string)
-  default     = []
-}
 
 # Tags
 variable "tags" {